requirements.txt
django
django-widget-tweaks
psycopg2
djangorestframework